compositor/meson.build: Install weston.ini to weston.ini.example
This would help out users have a base ini configuration file that they
can use directly without having to write one from scratch.

This install weston.ini.in to weston.ini.example to make it obvious this
is file that's *not* parsed at runtime but rather than needs to be
passed to Weston if one ones to customize stuff.

Fixes: #761
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>