drm-backend: Add an alias name mechanism for drm connectors
To address the issue of inconsistent DRM connector names on multi-card
devices, an alias name mechanism for DRM connectors is introduced. A udev
rule can be used to add a property WL_OUTPUT_ALIAS_NAME to the DRM
connector to represent its alias name. This way, the alias name can be
used in the input's udev rule to bind the output.

Signed-off-by: zhou,liang <liang.zhou@gehealthcare.com>