libweston: Make weston_compositor a singleton
libweston doesn't really support multiple compositor instances and it
isn't clear to me how to correctly support it. In the meantime, it
would be useful to retrieve the current unique instance for use in
other parts of the code like for instance from the logging and/or the
assertion subsystems.

This commit enforces a single allowed instanciation of the
weston_compositor structure and provides a new libweston-internal
symbol to retrieve its position in memory.

Signed-off-by: Loïc Molinari <loic.molinari@collabora.com>