build: Enable NDEBUG in release
New macros have been added in order to get rid of the warnings in
release builds when NDEBUG is defined. MAYBE_UNUSED() in helpers.h
marks a variable as unused when the assert in which it's used is
compiled out. WESTON_ASSERT_IF() and WESTON_DASSERT_IF() in
weston-assert.h prevent empty blocks when an assert is compiled out.

Signed-off-by: Loïc Molinari <loic.molinari@collabora.com>