shared: Enforce semicolons after asserts
Asserts are implemented in a way that allows the retrieval of the
condition value using assignment. This allows the assert unit test to
check for the result of the condition. This commit changes the assert
unit test using a custom handler so that it can set a boolean when the
condition is true.

This allows to get rid of the assignment use case in asserts and to
implement them using the more classical "do { ... } while (0)"
construct which has the advantage of enforcing semicolons at the end
of asserts at build time.

Signed-off-by: Loïc Molinari <loic.molinari@collabora.com>