backend-drm: do not disable underlay planes based on b->format
Since commit "backend-drm: improve code that chooses output->format", we
changed how b->format (the GBM format from the config file) is used.

There are config options in weston.ini that allow us to ignore the GBM
format set. So what matters to us is which output->format is being used,
and we compute that in drm_output_pick_format_egl() and
drm_output_pick_format_pixman().

Instead of disabling underlays when b->format does not have alpha
channel, disable that when output->format (for any of the outputs) has
no alpha channel.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>