input: Switch keyboard keymap layout on the fly
This introduces some helpers to handle a potential keyboard
keymap layout change, doing that dynamically upon a key code
change.

Requires having a udev rule matching a particular keyboard
with ENV{KEYBOARD_LAYOUT} set to a keymap layout (i.e., de, fr, es)
for a particular keyboard. As we don't pass udev attribs to ini file
this is the most accurate way of mapping a keyboard to a particular
layout.

When a key is pressed we determine if we need to switch the keymap
to that particular one, do not switch, or switch the default one (set in
the ini file). A cached keymap is also set-up and used to avoid
re-compiling each time as well as well as a default keymap.

Add small entry in man page about kEYBOARD_LAYOUT to notify users about
this feature.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>