clients/image: use embedded CICP colorimetry data to present image
For image files that have embedded CICP colorimetry data, use the CM&HDR
protocol extension to present taking that into account.

We use the CM&HDR extension to create a parametric image description
based on the colorimetry. Also through the CM&HDR extension we take this
image description and set it to be the surface image description of the
surface in which we present the image.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>