compositor: Fix up damage area of scaled surfaces
The compositor doesn't take scaling into account when setting up
damage areas. This can, in some cases, leave "dirty pixels" on the
renderbuffer. The pixels on the border of the damage area of a scaled
surface can spill over the area by n pixels because of bilinear
filtering, where n is scale / 2.

For example, a client with a bouncing quad might define 2 damage quads
each frame to update the new position (damage quad A) and to clean up
the previous one (damage quad B). If the surface is scaled, the pixels
at the border of damage quad A spill over damage quad B and remain
uncleaned on the renderbuffer.

This commit enlarges damage areas of scaled surfaces by the right
amount of pixels to prevent that.

Fixes #441

Signed-off-by: Loïc Molinari <loic.molinari@collabora.com>