libweston: Increase the cadence of the timer counter
Computing the frame rate counter is happening when the timer fires,
being set by default to a 1 second time interval.

The issue with that is we might loose quite a bit information/precision
until the next time the timer fires. With this change we increases the
firing cadence to 200 ms, thus giving us the possibility to count and
track more accurately what happens with the application.

While doing so we also compute the time in which the previous timer was
fired and use that delta instead of the default 200 ms when computing
the frame rate. All of these are aggregated over a 1 second period to
provide human-like number when calling the appropriate debug scope
(scene-graph, drm-backend).

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>