test-close-animation: Exercise closing of window on fade out animation
Two tests are added to exercise some code coverage on the destroying
path for a surface/view whilst having a fade out animation.

One test is verifying that the animation is indeed working by checking
the alpha value of a view created by desktop-shell.
If we time out here and animation never finishes meson should
time out the test as well and ultimately fail.

Another test similar to the previous one, is that we move out the window output
of the output which might trigger a crash.

Note that we deliberately don't (properly) cleanup the xdg_surface to
still have view mapped at destroy time as otherwise we won't hit that
desktop-shell removal part.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>