libweston: Send frame callbacks to occluded views
This fixes an issue where views added to hidden layer will no longer
be able to act upon the configure events sent out by the compositor.

With commit b2e6a6438f7, "libweston: Don't add frame callbacks from
occluded paint nodes" we made it so we do not send out frame callback if
the paint nodes are not visible. This has the side-effect of
affecting also paint nodes/views which were previously added into
hidden layers (by the shells).

This is a compromise to still allow clients still receive frame callbacks
whenever the shell decide that will move put that view into a hidden layer.

By default the normal behaviour will be kept (not sending frame
callbacks to occlduded views) and with this
'send-frame-callback-to-occluded-surface' will alter that behaviour.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>