backend: Add input support to headless backend
Currently, the headless backend doesn't support any kind of
input. This means that is not possible to send keystrokes to the
compositor or the applications running inside.

This patch allows to send keystrokes to the compositor from
STDIN. Just entering each key and a space, and ending the
sequence with a return. For example, typing

    super s

followed by RETURN, will send a Super+S keystroke, thus
launching a screen capture (which, for some reason, only works
with the PIXMAN rendered with the headless backend).

The modifiers (shift, ctrl, alt, super) remain pressed until a
carriage return is received. All other keys sent a PRESS event
followed by a RELEASE event.

The patch is still incomplete, because a lot of the keycodes are
missing, and also want to add an extra parameter to allow to
specified a named pipe instead of STDIN if desired.