libweston: move weston-assert from shared to include/libweston
In the next commits we'll integrate the weston-assert macros (not
test-assert) with our log infrastructure. In order to do that, restrict
such macros to libweston and its users.

test-assert macros do not take a compositor parameter and they log
directly to stderr, but weston-assert macros takes the compositor param
and the goal is to log that using our log infrastructure. We do that in
the next commits.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>