backend-drm: Add 'link bpc' connector property
The 'link bpc' DRM connector property can be used to retrieve the
physical bpc value.

This property can be used to find out the physical link value as
drivers will try to lower the bpc value when creaing an optimal output.

If we detect change we would notify the user if the value set with 'max
bpc' would be different than that of 'link bpc'.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>