compositor: Check output-capture for recording censor
Not all output capture activates disable-planes; notably, writeback
bypasses this. Check explicitly for output-capture tasks to activate
surface censoring.

This is not expected to impact any real-world users, as certified
implementations of protected-content usecases should already be taking
measures to ensure that the content cannot be exfiltrated even without
Weston doing the same.

Signed-off-by: Daniel Stone <daniels@collabora.com>