libweston: avoid bad recursion between assert and log infrastructure
Our custom weston_assert_* macros are now integrated with our log
infrastructure. But if the log infrastructure tries to use the custom
macros itself, that could lead to a bad recursion:

1. weston_assert_* macro is called on a failure, we try to log
2. bug in our log infrastructure may lead to a second weston_assert_*
   call
3. second weston_assert_* makes us to try to use the log infrastructure
   again...

To avoid this, add a new custom_assert_fail_ that uses vfprintf()
instead of weston_log_scope_vprintf(). With that, we can use the custom
weston_assert_* macros safely in the log infrastructure implementation.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>