tests: Handle transformed outputs in breakpoint helper
We have assert_output_matches() to make sure that the output we get in
our breakpoint is the one we wanted. Unfortunately, this didn't account
for the discrepancy in width and height between client and server when
the output was rotated.

Account for this so it can work with rotated outputs.

Signed-off-by: Daniel Stone <daniels@collabora.com>