tests: Don't use screenshots for output-damage
We have plenty of screenshot testing to ensure that our renderers are
doing the right thing. What output-damage wants to measure - that we're
only repainting what we've been asked to repaint and no more - is much
more easily done by directly testing the damage region handed to
repaint.

We can add more nuanced testing later on about specific rendering
methods to ensure we don't lose optimisations, but in the meantime this
not only speeds up output-damage testing, but also makes it a lot easier
to clean up some of our internals.

Signed-off-by: Daniel Stone <daniels@collabora.com>