content-protection: Consider recording to be effective disable
Rather than having to account for screen capture separately during
repaint, just force the output to be seen as unprotected when it is
activated.

This means that our condition for content censoring becomes much more
simple, and we know much closer to the source when those conditions have
changed, rather than just checking every repaint.

Signed-off-by: Daniel Stone <daniels@collabora.com>