paint-node: Consolidate early paint-node updates
We only need to rewrite the paint node's attributes if the paint node is
params-dirty (we have a different buffer type), or view-dirty (one of
many things, including an effective protection change, alpha change,
etc).

Consolidate this into a single spot we can use to deal with those
changes, and only change it when necessary.

Signed-off-by: Daniel Stone <daniels@collabora.com>