paint-node: Only update paint-node visibility when necessary
We know when the paint node visibility may change - because we've
moved or restacked the view, because its blend/opaque regions have
changed, or because the output has been changed. So we only need to
update the paint-node visible regions when this happens.

Previously, we had to do this every frame because content-protection
changes may have happened from under us. Now that we systematically
track this and mark things as dirty at the source, we don't need to
guess at repaint.

Signed-off-by: Daniel Stone <daniels@collabora.com>