libweston: split weston-assert.h and its implementation
Up to now, custom_assert_fail_ was defined in the same file in which we
wrote the weston-assert implementation (weston-assert.h), like this:

 #ifndef custom_assert_fail_
 #define custom_assert_fail_ weston_assert_fail_
 #endif

With this #ifned condition, other headers would be able to define their
own custom_assert_fail_ and import weston-assert.h. But this design is a
bit fragile, as it could do things "behind our back". E.g. if we have a
implementation that should be adding its own custom_assert_fail_ but
forgets, we would not know.

Instead, let's switch to a design that does not hide errors: split the
weston-assert implementation and the header that defines the
custom_assert_fail_, and avoid using the #ifndef condition. If somebody
messes with the ordering of #import's, duplicating custom_assert_fail_
or forgetting to add one, it will fail at build time.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>