shared: introduce assert macros without a compositor context parameter
weston-assert.h depends on a struct weston_compositor context parameter,
so we are able to log to the proper log scope. But we only made the
integration with the logging infrastructure recently, since commit
"libweston: integrate weston-assert with our log infrastructure".

Up to now we've been using these asserts on places where we shouldn't,
simply setting compositor to NULL (as it was unused). But now this
changes, so let's add a variant of the assert macros that can be used
without a compositor context parameter.

This should *not* be used inside libweston; it is intended for the test
suite, clients, etc, i.e. code that can't supply a valid compositor
context parameter.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>