tests: convert test suite to use assert_* macros
This converts the test suite to make use of the new assert_* macros,
completely replacing the test_assert_* macros.

This also changes our assert-test. It was testing weston_assert_* macros
with the compositor context parameter, but passing compositor == NULL
(which was awkward). As now weston-test-assert.h header imports the
assert_* macros, let's change assert-test to test these macros instead.

In the next commits we add a test for the weston_assert_* macros with
a proper compositor context parameter.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>